Game Room development

+++ monopoly/src/main/java/com/zzaip/monopoly/communication/game_room/GameRoom.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class GameRoom {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int gameRoomId;
+    private long gameRoomId;
 
     private String roomName;
